PS-4 Added separate component for contact form and changed layout of the contact component
